configs: Add ASUS M5A97 motherboard
This patch adds support for the ASUS M5A97 (not R2.0, the original)
motherboard.

The labels map exactly to the BIOS labels for consistency sake.

The CPU and VDDA voltage limits are the absolute minimum and maximum
values the motherboard regulators can supply. Users can always fine-tune
these to their own preference if so desired.

Found values where compared against bios values. The bios does have some
extra settable voltages, but those are not affecting the ignore sensors.
Maybe the Asus software has names/formulas for them, but as this is
Windows only software, not usable.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>