libsensors: Avoid non-symlink sysfs device entries
Devices under the MMC / SDIO bus have a "device" entry that is not a
symbolic link; instead it contains the device ID of the SDIO device.
This causes the sysfs parsing code in libsensors to error out, as it
expects all "device" entries be symbolic links to the underlying device.

Check for non-symlink entries before calling realpath() to avoid using
it as a valid device link.

Signed-off-by: Chen-Yu Tsai <wenst@google.com>